settingページのコンポーネント化・num run buildの修正
