+++ src/app/_components/Input.tsx
@@ -3,9 +3,15 @@ interface InputProps {
   value: string;
   onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
   placeholder: string;
+  disabled?: boolean;
 }
 
-const Input: React.FC<InputProps> = ({ value, onChange, placeholder }) => {
+const Input: React.FC<InputProps> = ({
+  value,
+  onChange,
+  placeholder,
+  disabled,
+}) => {
   return (
     <div>
       <input
@@ -14,6 +20,7 @@ const Input: React.FC<InputProps> = ({ value, onChange, placeholder }) => {
         onChange={onChange}
         placeholder={placeholder}
         className="border p-2 w-full mb-4"
+        disabled={disabled}
       />
     </div>
   );
